fix and refactor room link og metadata for previews
