+++ app/src/checkRoomOwner.js
@@ -1,16 +1,12 @@
 'use strict';
 
-const { createClient } = require('@supabase/supabase-js');
+const { getSupabaseClient } = require('./lib/supabase');
 require('dotenv').config();
-
-// Initialize Supabase client
-const supabase = createClient(
-  process.env.SUPABASE_URL || 'http://127.0.0.1:54321',
-  process.env.SUPABASE_ANON_KEY || ''
-)
 
 async function invokeCheckRoomOwner(roomName) {
   try {
+    const supabase = getSupabaseClient();
+    if (!supabase) return false;
     const { data } = await supabase.functions.invoke('check-room-owner', {
       body: { p_room_name: roomName }
     })
